requirements.txt
torch==2.5.1+cu124
einops==0.8.1
lightning==2.5.1.post0
matplotlib==3.10.0
numpy==1.26.3
pandas==2.2.3
pytorch-forecasting==1.3.0
pytorch-wavelets==1.3.0
PyYAML==6.0.2
reformer-pytorch==1.4.4
scikit-learn==1.6.1
scipy==1.15.2
torchmetrics==1.7.1
